File discovery (#95)
* Remove `language` Parameter & Add `FileInfo` Class

This commit introduces a bunch of breaking changes regarding the use of
the language variable as an instace variable and in constructors. This
will break the Runner, SourceRepository, and many metric classes.

* Change getAst to get_ast

* Discover Files in SourceRepository

* Fix Fall Through of If-stmts & Refactor SourceRepo + Runner Inits

* Discover and Exclude Files in SourceRepository without Language

* Add Tree-sitter Language Dict to source_repository

* Add parsers dictionary to source repository

* Fix Runner and Writer Tests After Rebase

* Fix Errors in All Metric Tests

Fix runtime errors introduced by FileInfo, the files instance variable,
and the src_root instance variable.

This does not fix the AssertationErrors, which are a result of each
metric test expecting each metric to measure only one language at a
time. However, it now will measure up to all three when possible.

* Fix Assertation of Outcome

Because SourceRepository now doesn't discriminate on language at all,
each metric reports the result for all languages. So tests have been
collected, and each expected result is the sum of the previously
expected results.

Location assertations are still broken.

* Fix Location Assertations in Metric Tests

* Run Black Formatter

* Use Constants for Lang + Encoding in Source Repo

* Update Supported Encodings based on Documentation

Python supports:
https://docs.python.org/3.11/library/codecs.html#standard-encodings

CPD supports:
- https://pmd.github.io/pmd/pmd_userdocs_cpd.html (see `--encoding`
  option)
- https://docs.oracle.com/javase/7/docs/api/java/nio/charset/Charset.html

chardet supports:
https://chardet.readthedocs.io/en/latest/supported-encodings.html

* Use assertCountEqual in Test Arg. Count

* Use assertCountEqual in All Locations Tests

assertCountEqual only cares that contents of iterables are identical, not the order.

* Use constants in Source Repo Dicts